Move the components shown on the screen
This will move the components that you see on the screen to the screen folder.
